pyproject.toml
[build-system]
requires = ["pip>=21.3", "setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }


[project]
name = "jaxman"
description = "Jax-based implementation for multi-agent navigation"
readme = "README.md"
license = { file = "LICENSE" }
version = "0.0.1"
keywords = [
    "navigation",
    "multi-agent navigation",
    "multi-agent reinforcement learning",
]
maintainers = [
    { name = "Hikaru Asano", email = "asano-hikaru19@g.ecc.u-tokyo.ac.jp" },
]
classifiers = ["Programming Language :: Python :: 3.9.*"]
requires-python = ">=3.9"
dependencies = [
    # "jaxmapp @ git+https://github.com/omron-sinicx/jaxmapp.git@be73e892d140ac04d15ba2781d936704ab6e4a6a",
    "gym==0.26.2",
    # "pettingzoo==1.22.1",
]

[project.optional-dependencies]
dev = [
    "pysen==0.10.1",
    "black==22.3.0",
    "flake8==4.0.1",
    "isort==5.10.1",
    "mypy==0.910",
    "pytest",
    "jax==0.3.13",
    "jaxlib==0.3.10",
    "numpy",
    "flax==0.4.1",
    "tensorflow-probability==0.16.0",
    "tensorflow",
    "tensorboardx",
    "matplotlib",
    # "seaborn",
    "scikit-image",
    "pymap2d",
    "hydra-core",
    "opencv-python",
    "ray",
    "moviepy",
    "bootstrapped",
]

[tool.pysen]
version = "0.10.1"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
py_version = "py39"

[[tool.pysen.lint.mypy_targets]]
paths = ["."]
